calcuate end date from current date

+++ dspace/modules/api/src/main/java/org/dspace/curate/CurationWeeklyReport.java
@@ -19,10 +19,10 @@ import java.util.List;
 import java.util.ArrayList;
 import java.util.Properties;
 import java.util.Date;
-import java.util.Calendar;
 import java.util.Scanner;
 import java.util.Calendar;
 import java.util.GregorianCalendar;
+import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.text.ParseException;
 
@@ -486,7 +486,7 @@ public class CurationWeeklyReport extends AbstractCurationTask {
        Get beginning date.
     **/
     
-	public Date getADate(byte numberOfDaysAgo)
+	public Date getADate(int numberOfDaysAgo)
 	{
         DateFormat format = new SimpleDateFormat("yyyyMMdd");
     	Calendar calcuatedDate = new GregorianCalendar();
